backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-JWT-Extended
Werkzeug
email
smtplib
